Fix claspignore for node_modules
