+++ tests/commands/push.ts
@@ -1,0 +1,41 @@
+import { expect } from 'chai';
+import * as fs from 'fs-extra';
+import { describe, it } from 'mocha';
+const { spawnSync } = require('child_process');
+
+import {
+  CLASP,
+  TEST_APPSSCRIPT_JSON,
+  TEST_CODE_JS,
+} from '../constants';
+
+import {
+  cleanup,
+  setup,
+} from '../functions';
+
+describe('Test clasp push function', () => {
+  before(setup);
+  it('should push local project correctly', () => {
+    fs.writeFileSync('Code.js', TEST_CODE_JS);
+    fs.writeFileSync('.claspignore', '**/**\n!Code.js\n!appsscript.json');
+    const result = spawnSync(
+      CLASP, ['push'], { encoding: 'utf8', input: 'y' },
+    );
+    expect(result.stdout).to.contain('Pushed');
+    expect(result.stdout).to.contain('files.');
+    expect(result.status).to.equal(0);
+  });
+  it.skip('should return non-0 exit code when push failed', () => {
+    fs.writeFileSync('.claspignore', '**/**\n!Code.js\n!appsscript.json\n!unexpected_file');
+    fs.writeFileSync('unexpected_file', TEST_CODE_JS);
+    const result = spawnSync(
+      CLASP, ['push'], { encoding: 'utf8', stdin: 'y'},
+    );
+    expect(result.stderr).to.contain('Invalid value at');
+    expect(result.stderr).to.contain('UNEXPECTED_FILE');
+    expect(result.stderr).to.contain('Files to push were:');
+    expect(result.status).to.equal(1);
+  });
+  // after(cleanup);
+});
